Add enqueue_many() with optimized insert query
Both enqueue() and enqueue_many() now use an optimized insert that can
handle any number of jobs (not using a list of VALUES(), so there isn't
a parameter limit for the binary postgres protocol).

Tests are a little cleaner as well due to the use of a mark to configure
the oban fixture in each test.