Normalize configuring stager interval
The stager now accepts options like the other core modules, through a
dict using the core name. That changes stager config from a top level
`stage_interval=_`, to a nested `stager={"interval": _}`. This is
consistent with how other core modules accept options, and it enables
direct configuration of the stager.