Implement testing.drain_queue() with async changes
This adds a helper for synchronously draining all jobs in a queue, using
options to control recursion, scheduling, and safety that mirror those
from Oban Elixir (though the defaults match Pro's drain_jobs).

While testing `with_recursion`, it became apparent that a worker's
`process()` method needed to be async. Otherwise, it's very difficult to
enqueue more jobs because it's an async operation. That had far-reaching
changes, but now all tests and documentation is updated for async
workers.